Improve "Volunteer Hours" page
Add the dropdown menu list to display the available activity/program that volunteers can perform, with the ability to customize their own activity name when clicking "Other". Also, adding the availability day of the week, date range and time that volunteers can do for the next 365 days from the current day that they register on the form.